requirements.txt
PySide6>=6.8
PyPDF2>=3.0
openpyxl>=3.1
Pillow>=10.0
pytesseract>=0.3
pdf2image>=1.17
